+++ packages/preprocess/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @evidence-dev/preprocess
 
+## 4.0.0-usql.0
+
+### Major Changes
+
+- cb0fc468: This update includes major changes to the way Evidence interacts with data.
+  Instead of running queries against the production database, and including it
+  with the project as pre-rendered, static JSON data; those queries are now stored as .parquet files.
+
+  .parquet enables the use of DuckDB on the client, allowing for much greater levels of interactivity
+  on pages, and interoperability between different data sources (e.g. joins across postgres & mysql).
+
+### Patch Changes
+
+- 75560a31: Consolidate tailwind presets into tailwind package
+
 ## 3.0.1
 
 ### Patch Changes