+++ sites/test-env/CHANGELOG.md
@@ -1,5 +1,33 @@
 # evidence-test-environment
 
+## 3.0.0-usql.1
+
+### Patch Changes
+
+- @evidence-dev/evidence@20.0.0-usql.1
+
+## 3.0.0-usql.0
+
+### Major Changes
+
+- cb0fc468: This update includes major changes to the way Evidence interacts with data.
+  Instead of running queries against the production database, and including it
+  with the project as pre-rendered, static JSON data; those queries are now stored as .parquet files.
+
+  .parquet enables the use of DuckDB on the client, allowing for much greater levels of interactivity
+  on pages, and interoperability between different data sources (e.g. joins across postgres & mysql).
+
+### Patch Changes
+
+- Updated dependencies [cb0fc468]
+- Updated dependencies [4e783f36]
+- Updated dependencies [75560a31]
+  - @evidence-dev/component-utilities@2.0.0-usql.0
+  - @evidence-dev/core-components@2.0.0-usql.0
+  - @evidence-dev/duckdb@1.0.0-usql.0
+  - @evidence-dev/evidence@20.0.0-usql.0
+  - @evidence-dev/faker-datasource@2.0.0-usql.0
+
 ## 2.0.16
 
 ### Patch Changes